+ accumulation of source compatible changes in 0.8 branch
git-svn-id: svn://svn.camaya.net/gloox/branches/0.8@1350 bb60cc1b-8fe5-0310-9879-aca5d778a633

+++ src/inbandbytestream.h
@@ -74,7 +74,7 @@ namespace gloox
        * Lets you retrieve this bytestream's block-size.
        * @return The bytestream's block-size.
        */
-      int blockSize() const { return m_blockSize; };
+      int blockSize() const { return (int)m_blockSize; };
 
       /**
        * Use this function to register an object that will receive any notifications from
